remove tslint rule
