+++ types/shopify-buy/index.d.ts
@@ -8,15 +8,53 @@
 //                 Martin Badin <https://github.com/martin-badin>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 
-import Client from './client';
+import GraphQLJSClient from './graphql-client';
+import Config from './config';
+import { Node } from './storefront/graphql';
+import CollectionResource from './collection-resource';
+import CheckoutResource from './checkout-resource';
+import ImageResource from './image-resource';
+import ProductResource from './product-resource';
+import ShopResource from './shop-resource';
 
-export * from './storefront/cart';
-export * from './storefront/common';
-export * from './storefront/customers';
-export * from './storefront/graphql';
-export * from './storefront/checkout';
-export * from './storefront/online_store';
-export * from './storefront/orders';
-export * from './storefront/products';
+import { Collection, Product, ProductVariant } from './storefront/products';
+import { Checkout } from './storefront/checkout';
+import { Shop } from './storefront/online_store';
 
-export default Client;
+declare namespace ShopifyBuy {
+    export { Product, ProductVariant, Collection, Checkout, Shop, Node };
+}
+
+interface FetchOptions {
+    body: string;
+    method: string;
+    mode: string;
+    headers: Record<string, string>;
+}
+
+type FetchFunction = (url: string, options: FetchOptions) => Promise<unknown>;
+
+declare class ShopifyBuy {
+    collection: CollectionResource;
+    checkout: CheckoutResource;
+    image: ImageResource;
+    product: ProductResource;
+    shop: ShopResource;
+
+    graphQLClient: GraphQLJSClient;
+
+    /**
+     * Primary entry point for building a new Client.
+     */
+    static buildClient(config: Config, fetchFunction?: FetchFunction): ShopifyBuy;
+
+    constructor(config: Config, GraphQLClientClass?: GraphQLJSClient, fetchFunction?: FetchFunction);
+
+    /**
+     * Fetches the next page of models
+     */
+    fetchNextPage<T extends Node>(models: T[]): Promise<T[]>;
+}
+
+export = ShopifyBuy;
+export as namespace ShopifyBuy;
